+++ juego/chess.py
@@ -2,6 +2,9 @@ from juego.board import Board
 from juego.exceptions import EmptyPosition
 from juego.exceptions import InvalidTurn
 from juego.exceptions import InvalidMove
+from juego.exceptions import InvalidMoveSameColor
+from juego.exceptions import InvalidCoordinate
+from juego.exceptions import InvalidSamePlace
 
 class Chess():
     def __init__(self):
@@ -10,21 +13,25 @@ class Chess():
 
     def is_playing(self):
         return True
-
-    def is_valid_coordiante(self, row, col):
-        return 0 <= row <= 8 and 0 <= col <= 8
     
     def move(self, from_row, from_col, to_row, to_col):
+        if not self.__board__.is_valid_coordinate(from_row, from_col) or not self.__board__.is_valid_coordinate(to_row, to_col):
+            raise InvalidCoordinate        
         piece = self.__board__.get_piece(from_row, from_col)
-        if not piece:
+        if not piece: 
             raise EmptyPosition()
         if not piece.get_color() == self.__turn__:
             raise InvalidTurn()
-        if not piece.valid_positions(from_row, from_col, to_row, to_col):
+        if not piece.valid_move(from_row, from_col, to_row, to_col):#Sin Test DUDA
             raise InvalidMove()
-        self.__board__.move(from_row, from_col, to_row, to_col)
-        self.change_turn()
-            
+        if from_row == to_row and from_col == to_col:
+            raise InvalidSamePlace
+        target_piece = self.__board__.get_piece(to_row, to_col)
+        if target_piece is not None and target_piece.get_color() == piece.get_color():
+            raise InvalidMoveSameColor()
+        self.__board__.move(from_row, from_col, to_row, to_col)#Sin Test
+        self.change_turn()#Sin Test
+        
     @property  
     def turn(self):
             return self.__turn__
@@ -34,9 +41,9 @@ class Chess():
 
     def change_turn(self):
         if self.__turn__ == "BLANCO":
-            self.__turn__ == "NEGRO"
+            self.__turn__ = "NEGRO"
         else:
-            self.__turn__ == "BLANCO"      
+            self.__turn__ = "BLANCO"