+++ changelog.md
@@ -5,9 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2024-09-30
+### Added 
+-Added new exceptions to chess and new tests to test_chess
+
+## [0.1.1] - 2024-09-29
+### Added 
+-Added new main
+
 ## [0.1.0] - 2024-09-28
 ### Added 
--Added new stylr to board
+-Added new style to board
 
 ## [0.0.29] - 2024-09-26
 ### Added